Started to convert gui_staticline_test to new structure, changed staticline hardware module to new structure, fixed NIDaq tutorial.
