+++ pylabnet/hardware/staticline/staticline.py
@@ -195,64 +195,3 @@ class StaticLineHardwareHandler():
                 self.log.info(
                     f"Setup of staticline {name} using module {module_name} successful."
                 )
-
-
-class Service(ServiceBase):
-
-    def exposed_up(self):
-        return self._module.up()
-
-    def exposed_down(self):
-        return self._module.down()
-
-    def exposed_get_name(self):
-        return self._module.get_name()
-
-
-class Client(ClientBase):
-
-    def up(self):
-        return self._service.exposed_up()
-
-    def down(self):
-        return self._service.exposed_down()
-
-    def get_name(self):
-        return self._service.exposed_get_name()
-
-
-def launch(**kwargs):
-    """ Connects to a NI DAQ as staticline
-
-    :param kwargs: (dict) containing relevant kwargs
-        :logger: instance of LogClient for logging purposes
-        :port: (int) port number for the Cnt Monitor server
-    """
-
-    import pylabnet.hardware.ni_daqs.nidaqmx_card as nidaqmx
-
-    device_name = 'cDAQ1Mod1'
-    daq = nidaqmx.Driver(device_name=device_name, logger=kwargs['logger'])
-
-    test_staticline = Driver(
-        name='Test staticline',
-        logger=kwargs['logger'],
-        hardware_module=daq,
-        ao_output='ao2',
-
-        down_voltage=0,
-        up_voltage=3.3,
-    )
-
-    # Instantiate Server
-    # Staticline server
-    staticline_service = Service()
-    staticline_service.assign_module(module=test_staticline)
-    staticline_service.assign_logger(logger=kwargs['logger'])
-    staticline_service_server = GenericServer(
-        service=staticline_service,
-        host='localhost',
-        port=kwargs['port']
-    )
-
-    staticline_service_server.start()
