+++ pylabnet/launchers/launch_control.py
@@ -75,9 +75,9 @@ class Controller:
             self.log_port = int(input('Please enter the master Logger Port:\n>> '))
             self.gui_port = int(input('Please enter the master GUI Port:\n>> '))
             hide_console()
-        
+
         sys.stdout = StringIO()
-        
+
         # Instantiate GUI application
         self.app = QtWidgets.QApplication(sys.argv)
         self.main_window = Window(self.app, gui_template=self.LOGGER_UI)
@@ -107,7 +107,7 @@ class Controller:
                 self.gui_logger.error(f'Failed to connect to GUI Server with IP address: {self.host}, '
                                       f'Port: {self:gui_port}')
                 raise
-            
+
             # Now update GUI to mirror clients
             self._copy_master()
 
@@ -122,8 +122,8 @@ class Controller:
             self.gui_service.assign_logger(logger=self.gui_logger)
             if self.gui_port is None:
                 self.gui_server, self.gui_port = create_server(
-                    self.gui_service, 
-                    logger=self.gui_logger, 
+                    self.gui_service,
+                    logger=self.gui_logger,
                     host=socket.gethostbyname(socket.gethostname())
                     )
             else:
@@ -218,21 +218,21 @@ class Controller:
         self.log_service = LogService()
         if self.LOG_PORT is None:
             self.log_server, self.log_port = create_server(
-                self.log_service, 
+                self.log_service,
                 host=socket.gethostbyname(socket.gethostname())
                 )
         else:
             try:
                 self.log_server = GenericServer(
                     service=self.log_service,
-                    host=socket.gethostname(socket.gethostname()), 
+                    host=socket.gethostname(socket.gethostname()),
                     port=self.LOG_PORT
                     )
             except ConnectionRefusedError:
                 print(f'Failed to insantiate Log Server at port {self.LOG_PORT}')
                 raise
         self.log_server.start()
-    
+
     def initialize_gui(self):
         """ Initializes basic GUI display """
 
@@ -268,10 +268,10 @@ class Controller:
 
     def update_proxy(self):
         """ Updates the proxy with new content using the buffer terminal"""
-        
+
         # Check clients and update
         self._pull_connections()
-        
+
         # Get buffer terminal
         buffer_terminal = self.gui_client.get_text('buffer')
 
@@ -280,7 +280,7 @@ class Controller:
 
         # Check if this failed
         if new_msg is '':
-            
+
             # Check if the buffer is ahead of our last update
             up_str = re.findall(r'!~\d+~!', new_msg)
             if len(up_str) > 0:
@@ -297,7 +297,7 @@ class Controller:
             except TypeError:
                 pass
             self.update_index = int(re.findall(r'\d+', re.findall(r'!~\d+~!', new_msg)[-1])[0])
-    
+
     def _configure_clicks(self):
         """ Configures what to do if script is clicked """
 
@@ -309,6 +309,16 @@ class Controller:
         Opens a new commandline subprocess using subprocess.Popen(bash_cmd) which runs the
         relevant python script, passing all relevant LogClient information via the commandline
         """
+
+        import ptvsd
+        import os
+        # 5678 is the default attach port in the VS Code debug configurations
+        #self.logger.info(f"Waiting for debugger attach to PID {os.getpid()}")
+        #self.logger.info(f"Waiting for debugger attach to PID {os.getpid()} (launcher_script)")
+        ptvsd.enable_attach(address=('localhost', 5678))
+        ptvsd.wait_for_attach()
+        breakpoint()
+
 
         script_to_run = list(self.script_list.keys())[list(self.script_list.values()).index(
             self.main_window.script_list.currentItem()
@@ -419,7 +429,6 @@ class Controller:
 
 def main():
     """ Runs the launch controller """
-
 
     log_controller = Controller()
